requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
sqlalchemy
pgvector
psycopg[binary]
pypdf
google-generativeai